requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
Deprecated==1.2.18
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
rich==13.5.2
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
semgrep==1.122.0     
pydantic==2.11.6
pydantic_core==2.33.2
openai==1.88.0
flask==3.0.3
flask-cors==6.0.1
pyfiglet==1.0.3
